PackageCuration: Move split command to separate sub-sub-command
To organise the curation helper commands better, as more commands will
be added in a later commit.

Renamed from `split-curations` as the curations context is now already
given from the sub-command.
