+++ helper-cli/src/main/kotlin/HelperMain.kt
@@ -42,11 +42,11 @@ import org.ossreviewtoolkit.helper.commands.ListStoredScanResultsCommand
 import org.ossreviewtoolkit.helper.commands.MapCopyrightsCommand
 import org.ossreviewtoolkit.helper.commands.MergeRepositoryConfigurationsCommand
 import org.ossreviewtoolkit.helper.commands.SetLabelsCommand
-import org.ossreviewtoolkit.helper.commands.SplitCurationsCommand
 import org.ossreviewtoolkit.helper.commands.SubtractScanResultsCommand
 import org.ossreviewtoolkit.helper.commands.TransformResultCommand
 import org.ossreviewtoolkit.helper.commands.VerifySourceArtifactCurationsCommand
 import org.ossreviewtoolkit.helper.commands.packageconfig.PackageConfigurationCommand
+import org.ossreviewtoolkit.helper.commands.packagecuration.PackageCurationCommand
 import org.ossreviewtoolkit.helper.commands.repoconfig.RepositoryConfigurationCommand
 import org.ossreviewtoolkit.helper.common.ORTH_NAME
 import org.ossreviewtoolkit.utils.printStackTrace
@@ -84,9 +84,9 @@ internal class HelperMain : CliktCommand(name = ORTH_NAME, epilog = "* denotes r
             MapCopyrightsCommand(),
             MergeRepositoryConfigurationsCommand(),
             PackageConfigurationCommand(),
+            PackageCurationCommand(),
             RepositoryConfigurationCommand(),
             SetLabelsCommand(),
-            SplitCurationsCommand(),
             SubtractScanResultsCommand(),
             TransformResultCommand(),
             VerifySourceArtifactCurationsCommand()
